Refactored active course
