+++ LanguageSchoolApp/LanguageSchoolApp/data/CourseApplications.json
@@ -12,5 +12,19 @@
     "CourseId": 1291962207,
     "Paused": false,
     "Acceptance": 0
+  },
+  "723148165": {
+    "Id": 723148165,
+    "StudentId": "student@gmail.com",
+    "CourseId": 1,
+    "Paused": false,
+    "Acceptance": 0
+  },
+  "1792788136": {
+    "Id": 1792788136,
+    "StudentId": "student@gmail.com",
+    "CourseId": 364486381,
+    "Paused": false,
+    "Acceptance": 0
   }
 }
