+++ LanguageSchoolApp/LanguageSchoolApp/data/Students.json
@@ -33,7 +33,7 @@
         }
       }
     ],
-    "EnrolledCourseId": -1,
+    "EnrolledCourseId": 952824467,
     "PenaltyPoints": [],
     "Name": "Aleksandar",
     "Surname": "Markovic",
