+++ LanguageSchoolApp/LanguageSchoolApp/data/Teachers.json
@@ -8,7 +8,11 @@
       2,
       10,
       5,
-      10
+      10,
+      2,
+      5,
+      9,
+      7
     ],
     "MyCoursesIds": [
       952824467,
