+++ LanguageSchoolApp/LanguageSchoolApp/view/Courses/ActiveCourseDetailsView.xaml
@@ -1,0 +1,76 @@
+﻿<UserControl x:Class="LanguageSchoolApp.view.Courses.ActiveCourseDetailsView"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:LanguageSchoolApp.view.Courses"
+             mc:Ignorable="d" 
+             d:DesignHeight="217" d:DesignWidth="279">
+    <Border CornerRadius="20"
+        Background="#fff"
+        BorderThickness="1">
+        <Border.BorderBrush>
+            <SolidColorBrush Color="#000" Opacity="0.10"/>
+        </Border.BorderBrush>
+        <Border.Effect>
+            <DropShadowEffect Color="#000"
+          Direction="315"
+          ShadowDepth="5"
+          BlurRadius="10"
+          Opacity="0.5"/>
+        </Border.Effect>
+        <Grid>
+            <Grid.RowDefinitions>
+                <RowDefinition/>
+                <RowDefinition Height="Auto"/>
+                <RowDefinition/>
+            </Grid.RowDefinitions>
+            <Label Content="{Binding LanguageProficiency}"
+               Margin="14,19,0,0"
+               FontSize="24"
+               FontFamily="Cambria"
+               FontWeight="Bold"
+               Grid.Row="0"/>
+            <WrapPanel Orientation="Vertical"
+                   Grid.Row="1"
+                   Margin="14,0,0,0">
+                <WrapPanel Orientation="Horizontal">
+                    <Label Content="Beginning Date:"
+                       FontSize="15"
+                       FontFamily="Cambria"/>
+                    <Label Content="{Binding BeginningDate, StringFormat=\{0:dd.MM.yyyy\}}"
+                       FontSize="15"
+                       FontFamily="Cambria"/>
+                </WrapPanel>
+                <WrapPanel Orientation="Horizontal">
+                    <Label Content="Duration:"
+                       FontSize="15"
+                       FontFamily="Cambria"/>
+                    <Label Content="{Binding Duration}"
+                       FontSize="15"
+                       FontFamily="Cambria"/>
+                </WrapPanel>
+                <WrapPanel Orientation="Horizontal">
+                    <Label Content="Course Type:"
+                       FontSize="15"
+                       FontFamily="Cambria"/>
+                    <Label Content="{Binding CourseType}"
+                       FontSize="15"
+                       FontFamily="Cambria"/>
+                </WrapPanel>
+            </WrapPanel>
+            <Button Style="{StaticResource RoundedButtonStyle}"
+                Content="Drop out"
+                Command="{Binding DropoutCommand}"
+                HorizontalAlignment="Right"
+                Margin="0,13,29,22"
+                Background="red"
+                Foreground="#fff"
+                FontFamily="Cambria"
+                FontSize="12"
+                Padding="20,0,20,0"
+                Grid.Row="2"
+                FontWeight="SemiBold"/>
+        </Grid>
+    </Border>
+</UserControl>
