+++ LanguageSchoolApp/LanguageSchoolApp/view/Courses/ActiveCoursePenaltyPointsView.xaml
@@ -1,0 +1,47 @@
+﻿<UserControl x:Class="LanguageSchoolApp.view.Courses.ActiveCoursePenaltyPointsView"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:LanguageSchoolApp.view.Courses"
+             mc:Ignorable="d" 
+             d:DesignHeight="193" d:DesignWidth="164">
+    <Border CornerRadius="20"
+        Background="#fff"
+        BorderThickness="1">
+        <Border.BorderBrush>
+            <SolidColorBrush Color="#000" Opacity="0.10"/>
+        </Border.BorderBrush>
+        <Border.Effect>
+            <DropShadowEffect Color="#000"
+                       Direction="315"
+                       ShadowDepth="5"
+                       BlurRadius="10"
+                       Opacity="0.5"/>
+        </Border.Effect>
+        <Grid>
+            <Grid.RowDefinitions>
+                <RowDefinition Height="Auto"/>
+                <RowDefinition Height="*"/>
+            </Grid.RowDefinitions>
+            <Label Content="Penalty Points" FontFamily="Cambria" FontSize="15" HorizontalAlignment="Center"
+               Margin="0,10,0,0"/>
+            <Border BorderBrush="#000" Width="112"
+                HorizontalAlignment="Center"
+                BorderThickness="0,0,0,1"/>
+            <ItemsControl Grid.Row="1" Margin="14,17,14,10" ItemsSource="{Binding PenaltyPoints}"
+                      BorderThickness="0" Background="Transparent">
+                <ItemsControl.ItemsPanel>
+                    <ItemsPanelTemplate>
+                        <WrapPanel/>
+                    </ItemsPanelTemplate>
+                </ItemsControl.ItemsPanel>
+                <ItemsControl.ItemTemplate>
+                    <DataTemplate>
+                        <Label Content="{Binding Reason}" HorizontalAlignment="Center" FontFamily="Cambria" FontSize="12"/>
+                    </DataTemplate>
+                </ItemsControl.ItemTemplate>
+            </ItemsControl>
+        </Grid>
+    </Border>
+</UserControl>
