+++ LanguageSchoolApp/LanguageSchoolApp/view/Courses/ActiveCourseTeacherDetailsView.xaml
@@ -1,0 +1,111 @@
+﻿<UserControl x:Class="LanguageSchoolApp.view.Courses.ActiveCourseTeacherDetailsView"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:LanguageSchoolApp.view.Courses"
+             mc:Ignorable="d" 
+             d:DesignHeight="217" d:DesignWidth="289">
+    <Border CornerRadius="20"
+        Background="#fff"
+        BorderThickness="1">
+        <Border.BorderBrush>
+            <SolidColorBrush Color="#000" Opacity="0.10"/>
+        </Border.BorderBrush>
+        <Border.Effect>
+            <DropShadowEffect Color="#000"
+          Direction="315"
+          ShadowDepth="5"
+          BlurRadius="10"
+          Opacity="0.5"/>
+        </Border.Effect>
+        <Grid>
+            <Grid.RowDefinitions>
+                <RowDefinition Height="Auto"/>
+                <RowDefinition Height="Auto"/>
+                <RowDefinition/>
+            </Grid.RowDefinitions>
+            <Label Content="Teacher"
+               FontFamily="Cambria"
+               FontSize="20"
+               FontWeight="Bold"
+               HorizontalAlignment="Center"
+               Margin="0,20,0,0"/>
+            <WrapPanel Orientation="Vertical" Margin="23,5,0,0" Grid.Row="1">
+                <WrapPanel Orientation="Horizontal">
+                    <Label Content="Name:"
+                       FontFamily="Cambria"
+                       FontSize="15"/>
+                    <Label Content="{Binding TeacherName}"
+                       FontFamily="Cambria"
+                       FontSize="15"/>
+                </WrapPanel>
+                <WrapPanel Orientation="Horizontal">
+                    <Label Content="Surname:"
+                       FontFamily="Cambria"
+                       FontSize="15"/>
+                    <Label Content="{Binding TeacherSurname}"
+                       FontFamily="Cambria"
+                       FontSize="15"/>
+                </WrapPanel>
+                <WrapPanel Orientation="Horizontal">
+                    <Label Content="Average Grade:"
+                       FontFamily="Cambria"
+                       FontSize="15"/>
+                    <Label Content="{Binding AverageGrade}"
+                       FontFamily="Cambria"
+                       FontSize="15"/>
+                </WrapPanel>
+            </WrapPanel>
+            <Grid Grid.Row="2">
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition Width="*"/>
+                    <ColumnDefinition Width="Auto"/>
+                </Grid.ColumnDefinitions>
+                <Grid Margin="23,25,5,0">
+                    <Grid.RowDefinitions>
+                        <RowDefinition Height="Auto"/>
+                        <RowDefinition Height="*"/>
+                    </Grid.RowDefinitions>
+                    <Grid>
+                        <Grid.ColumnDefinitions>
+                            <ColumnDefinition Width="Auto"/>
+                            <ColumnDefinition Width="Auto"/>
+                        </Grid.ColumnDefinitions>
+                        <TextBlock Grid.Column="0"
+                                   Text="Grade Teacher:"
+                                   FontFamily="Cambria"
+                                   FontSize="12"
+                                   VerticalAlignment="Center"/>
+                        <TextBox x:Name="TeachersGradeTextBox"
+                             Grid.Column="1"
+                             Text="{Binding TeachersGrade}"
+                             Height="14"
+                             Width="24"
+                             VerticalContentAlignment="Bottom"
+                             HorizontalContentAlignment="Right"
+                             BorderThickness="0"
+                             Background="Transparent"
+                             VerticalAlignment="Center"/>
+                    </Grid>
+                    <Border BorderBrush="#000" Width="101"
+                        HorizontalAlignment="Left"
+                        BorderThickness="0,0,0,1" Margin="0,4,0,0"/>
+                </Grid>
+
+                <Button x:Name="GradeTeacherButton"
+                    Content="Confirm"
+                    Command="{Binding GradeTeacherCommand}"
+                    Style="{StaticResource RoundedButtonStyle}"
+                    Background="#2E6C7B"
+                    Foreground="#fff"
+                    FontFamily="Cambria"
+                    HorizontalAlignment="Left"
+                    FontSize="12"
+                    Height="34"
+                    Padding="20,0,20,0"
+                    Margin="185,14,0,25"/>
+            </Grid>
+        </Grid>
+    </Border>
+</UserControl>
